chore: rename enum.py to types.py
